Realised (again) that extreme carbon was wrong
Although I was using fixed timesteps for the Vern8 algorithm (for the built in ODE solver) the intermediate timesteps that it calcualates are not regular and so the extreme carbon limit forcing was ~8 times too large. Now changed to RK4 (not the best but the highest order fixed intermediary timestep algorithm I think they have). I might impliment a better algorithm myself so that the losses can be directly implimented instead of through the derivitives.