First attempt, does not work
I have implimented using FunctionCallingCallback and handels the limit far worse than the previous implimentation and lets C go negative.

It does initially stop c at C_min but does not stablly keep it there. I have not yet diagnosed the cause but suspect it is because the function is called every timestep, not during the substeps.